chore(compass-component): add a separate component for spin loader with label
